Removed stale code, added BidirectionalMap.
The Bidirectional Map implements what was supposed to be done by the
stale code. It is needed to produce an efficient lookup from users to
their list of interests, and similarly, from a (host name, job name)
pair to the list of interested users.

It has not yet been connected to the functionality, since we have no way
to actually tell a user what jobs are available or even get a list of
users.